==> notebooks/01_data_handler.ipynb <==
import sys
sys.path.append('..')
%load_ext autoreload
%autoreload 2
%matplotlib inline

import os

from quick_pp.database.objects import Project
from quick_pp.database.db_connector import DBConnector

db_conn = DBConnector()
db_conn.setup_db()

example = 'carbonate'  #  'clastic'  #
folder = 'VOLVE' if example == 'clastic' else 'COSTA'
project_name = "MOCK_" + example

data_path = os.path.join('data', '01_raw', folder)
filenames = []
for root, dirs, files in os.walk(data_path):
    for file in files:
        filenames.append(os.path.join(root, file)) if file.endswith('.las') and not 'seismic' in root  else None
print(filenames)

with db_conn.get_session() as db_session:
    proj = Project(db_session, name=project_name)
    proj.read_las(filenames, depth_uom='m')

with db_conn.get_session() as db_session:
    proj = Project(db_session, name=project_name)
    df = proj.get_all_data()

df[df.WELL_NAME == 'HW-9'].head()

####  Steps to create "MOCK" project
1. Run the next cell.
2. Specify the example either clastic or carbonate.
3. MOCK.qppp project will be saved in notebooks\data\04_project folder.

* Note that the required curves in the LAS files are 'GR', 'RT', 'NPHI', 'RHOB'

# example = 'clastic'  # 'carbonate'  #  
# folder = 'VOLVE2' if example == 'clastic' else 'COSTA'

# data_path = os.path.join('data', '01_raw', folder)
# filenames = []
# for root, dirs, files in os.walk(data_path):
#     for file in files:
#         filenames.append(os.path.join(root, file)) if file.endswith('.las') and not 'seismic' in root  else None
# print(filenames)

# project_name = "MOCK_" + example
# project = Project(name=project_name)
# project.read_las(filenames)
# project.save()

# # clear_output()               

# from tkinter import Tk, filedialog

# root = Tk()
# file_objects = filedialog.askopenfiles(title='Choose well Log ASCII Standard (LAS) files to be combined',
#                                         filetype=(('LAS Files', '*.LAS *.las'), ('All Files', '*.*')),
#                                         multiple=True,
#                                         mode='rb')
# root.destroy()
# if file_objects:
#     project_name = "MOCK"    
#     project = Project(name=project_name)
#     project.read_las([f.name for f in file_objects])
#     project.save()

# clear_output()

# import pandas as pd

# df = pd.read_parquet(r"xxx.parquet")
# marker_df = pd.read_csv(r"zzz.csv")
# marker_df['WELL_NAME'] = marker_df['Well identifier (UWI)']
# marker_df['DEPTH'] = marker_df['MDDF(ft)']
# marker_df['ZONES'] = marker_df['Surface']

# return_df = pd.DataFrame()
# for well, data in df.drop('ZONES', axis=1).groupby('WELL_NAME'):
#     temp_df = marker_df[marker_df.WELL_NAME == well][['DEPTH', 'ZONES']].copy().sort_values('DEPTH')
#     temp_df = pd.merge_asof(
#         data, temp_df, on='DEPTH', direction='forward', tolerance=50
#     )
#     return_df = pd.concat([return_df, temp_df])

# project_name = "BEKK_CLEAN"    
# project = Project(name=project_name)
# project.update_data(return_df)
# project.save()

# # Loading LAS files

# root = Tk()
# file_objects = filedialog.askopenfiles(title='Choose well Log ASCII Standard (LAS) files to be combined',
#                                         filetype=(('LAS Files', '*.LAS *.las'), ('All Files', '*.*')),
#                                         multiple=True,
#                                         mode='rb')
# root.destroy()
# input_df, _ = las.read_las_files(file_objects)

# # Loading LAS files

# root = Tk()
# file_objects = filedialog.askopenfiles(title='Choose well Log ASCII Standard (LAS) files to be combined',
#                                         filetype=(('LAS Files', '*.LAS *.las'), ('All Files', '*.*')),
#                                         multiple=True,
#                                         mode='rb')
# root.destroy()
# output_df, _ = las.read_las_files(file_objects)

vars_units = dict(
    DRHO='g/cc',
    DT='us/ft',
    DTS='us/ft',
    RHOB='g/cc',
    NPHI='v/v',
    GR='GAPI',
    RT='ohm.m',
    RM='ohm.m',
    RS='ohm.m',
    BVW='v/v',
    KLOGH='mD',
    PHIF='v/v',
    SAND='v/v',
    VCARB='v/v',
    VSH='v/v',
    PEF='b/elec',
    SW='v/v',
    CALI='inches',
    BS='inches',
    ROP='m/hr'
)


# output_folder = r'data\01_raw\VOLVE'

# merged_df = pd.DataFrame()

# for well, data in input_df.groupby('WELL_NAME'):
#     temp_output_df = output_df[output_df['WELL_NAME'] == well].sort_values(by='DEPTH')
#     temp_df = pd.merge_asof(data.sort_values(by='DEPTH'), temp_output_df, on=['DEPTH'], suffixes=('', '_output'))
#     # Get columns with _output suffix
#     output_cols = [col for col in temp_df.columns if col.endswith('_output')]
    
#     # For each _output column, fill null values in original column with values from _output
#     for col in output_cols:
#         base_col = col.replace('_output', '')
#         temp_df[base_col].fillna(temp_df[col], inplace=True)
#         # Drop the _output column
#         temp_df.drop(columns=[col], inplace=True)

#     temp_df.drop(columns=['WELL_NAME', 'UWI'], inplace=True)
#     temp_df.dropna(how='all', axis=1, inplace=True)
#     las.export_to_las(temp_df, folder=output_folder, well_name=f'{well}', vars_units=vars_units)

#     merged_df = pd.concat([merged_df, temp_df])

# # merged_df.to_csv('merged_df.csv', index=False)

# # Loading DLIS files
# from dlisio import dlis

# root = Tk()
# file_objects = filedialog.askopenfiles(title='Choose well Log ASCII Standard (LAS) files to be combined',
#                                         filetype=(('LAS Files', '*.DLIS *.dlis'), ('All Files', '*.*')),
#                                         multiple=True,
#                                         mode='rb')
# root.destroy()

# dlis_input = pd.DataFrame()
# for file_object in file_objects:
#     with dlis.load(file_object.name) as (file, *tail):
#         well_name = file.origins[0].well_name
#         well_name = well_name.replace('/', '-').replace(' ', '-')
#         for fr in file.frames:
#             temp_df = pd.DataFrame(fr.curves())
#             temp_df['WELL_NAME'] = well_name
#             temp_df.drop(columns=['FRAMENO'], inplace=True)
#             dlis_input = pd.concat([dlis_input, temp_df], ignore_index=True)

# # Need to convert depth from 0.1 inch to 0.5 ft and to metres
# dlis_input['DEPTH'] = dlis_input.DEPTH / 120 / 3.281

# # Loading DLIS files
# from dlisio import dlis

# root = Tk()
# file_objects = filedialog.askopenfiles(title='Choose well Log ASCII Standard (LAS) files to be combined',
#                                         filetype=(('LAS Files', '*.DLIS *.dlis'), ('All Files', '*.*')),
#                                         multiple=True,
#                                         mode='rb')
# root.destroy()

# dlis_output = pd.DataFrame()
# for file_object in file_objects:
#     with dlis.load(file_object.name) as (file, *tail):
#         well_name = file.origins[0].well_name
#         well_name = well_name.replace('/', '-').replace(' ', '-')
#         for fr in file.frames:
#             temp_df = pd.DataFrame(fr.curves())
#             temp_df['WELL_NAME'] = well_name
#             temp_df.drop(columns=['FRAMENO'], inplace=True)
#             dlis_output = pd.concat([dlis_output, temp_df], ignore_index=True)

# # Need to convert depth from 0.1 inch to 0.5 ft 
# dlis_output['DEPTH'] = dlis_output.DEPTH / 120 / 3.281

# output_folder = r'data\01_raw\VOLVE'

# dlis_merged_df = pd.DataFrame()

# for well, data in dlis_input.groupby('WELL_NAME'):
#     temp_output_df = dlis_output[dlis_output['WELL_NAME'] == well].sort_values(by='DEPTH')
#     temp_df = pd.merge_asof(data.sort_values(by='DEPTH'), temp_output_df, on=['DEPTH'], suffixes=('', '_output'))
#     # Get columns with _output suffix
#     output_cols = [col for col in temp_df.columns if col.endswith('_output')]
    
#     # For each _output column, fill null values in original column with values from _output
#     for col in output_cols:
#         base_col = col.replace('_output', '')
#         temp_df[base_col].fillna(temp_df[col], inplace=True)
#         # Drop the _output column
#         temp_df.drop(columns=[col], inplace=True)

#     temp_df.drop(columns=['WELL_NAME', 'UWI'], inplace=True, errors='ignore')
#     temp_df.dropna(how='all', axis=1, inplace=True)
#     las.export_to_las(temp_df, folder=output_folder, well_name=f'{well.replace("/", "-").replace(" ", "-")}', vars_units=vars_units)

#     dlis_merged_df = pd.concat([dlis_merged_df, temp_df])

# # dlis_merged_df.to_csv('dlis_merged_df.csv', index=False)
